#165486106 Migrate transfers endpoint from file system to the app database (#54)
* feature(get user): integrate users GET endpoint
Migrate function for getting users to database.
Database supports getting a single user or getting
all registered users.
[Finishes #165448090]

* feature(update user): migrate update user to DB
Migrate user function that previously relied on
the file system to the database.
[Finishes #165475158]

* feature(integrate delete):migrate delete function
Migrate the user delete function to the database
ensuring that the function works even better now
and relevant tests are passed.
[Finishes #165478469]

* feature(integrate delete):migrate delete function
Migrate the user delete function to the database
ensuring that the function works even better now
and relevant tests are passed.
[Finishes #165478469]

* feature(account):migrate create account to DB
Migrate function for creating new banka account
to the Database. Ensure all tests pass
[Delivers #165479889]

* feature(account):migrate create account to DB
Migrate function for creating new banka account
to the Database. Ensure all tests pass
[Delivers #165479889]

* feature(patch):migrate PATCH account logic to DB
Migrate function for activating and deactivating
user accounts to the database.
[Delivers #165482120]

* feature(delete): migrate delete account to DB
Delete account function to interact with database
Test for 500 error by introducing stubs
[Delivers #165482970]

* feature(transactions): migrate transactions to DB
Migrate transactions endpoint to the database
from the file system.
Refactor error and email messages.
[Delivers #165483205]

* feature(transfers): migrate transfers to db
Migrate transfers endpoint and controllers to
interact with the database and not the file
system. Ensure tests are passing.
[Delivers #165486106]

* #165409881 Get client account details from the app database (#55)

* feature(account): get client account details
There was no function that allowed users to see
their account details or staff to see the details
on a user account. This commit provides for it.
[Delivers #165409881]

* #165491005 Get all accounts saved to the app database (#56)

* feature(accounts): build GET accounts endpoint
Staff and admin can make requests and get an
array containing all accounts.
[Delivers #165491005]

* feature(accounts): get accounts by status (#57)

Build controllers for getting accounts by status
Enable support for getting active and dormant accounts
[Finishes #165491418]